changed the way the button to launch the console manager is rendered and changed how new notifications visually alert the user, and how those alerts get cleared
